project/line/src/requirements.txt
line-bot-sdk==1.8.0
Flask==1.0.2
gspread==3.1.0
oauth2client==4.1.3
itsdangerous==2.0.1
jinja2==3.0.3
Werkzeug~=2.0.0